Editar entidades

+++ app/Http/Controllers/DistribuicaoController.php
@@ -44,55 +44,96 @@ class DistribuicaoController extends Controller
 
   public function remover(Request $request){
       $distribuicao = \App\Distribuicao::find($request->id);
-      $distribuicao->delete();
-      session()->flash('success', 'Distribuicao removida com sucesso.');
+
+      if (isset($distribuicao)) {
+        $distribuicao->delete();
+        session()->flash('success', 'Distribuicao removida com sucesso.');
+        return redirect()->route('/distribuicao/listar');
+      }
+      
+      session()->flash('success', 'Distribuicao não existe.');
       return redirect()->route('/distribuicao/listar');
     }
 
   public function editar(Request $request){
       $distribuicao = \App\Distribuicao::find($request->id);
-      $escolas = \App\Escola::all();
-      return view("EditarDistribuicao", [
-          "distribuicao" => $distribuicao,
-          "escolas" => $escolas,
-      ]);
+      if (isset($distribuicao)) {
+        $escolas = \App\Escola::all();
+        return view("EditarDistribuicao", [
+            "distribuicao" => $distribuicao,
+            "escolas" => $escolas,
+        ]);
+      }
+      session()->flash('success', 'Distribuicao não existe.');
+      return redirect()->route('/distribuicao/listar');
   }
 
   public function salvar(Request $request){
       $distribuicao = \App\Distribuicao::find($request->id);
 
-      $distribuicao->observacao = $request->observacao;
-      $distribuicao->escola_id = $request->escola_id;
-      $distribuicao->save();
+      if (isset($distribuicao)) {
+        $distribuicao->observacao = $request->observacao;
+        $distribuicao->escola_id = $request->escola_id;
+        $distribuicao->save();
 
-      session()->flash('success', 'Distribuicao modificada com sucesso.');
+        session()->flash('success', 'Distribuicao modificada com sucesso.');
+        return redirect()->route('/distribuicao/listar');
+      }
+      
+      session()->flash('success', 'Distribuicao não existe.');
       return redirect()->route('/distribuicao/listar');
   }
 
   public function inserirItemDistribuicao(Request $request) {
-    $distribuicao_item = new \App\Distribuicao_item();
-    $distribuicao_item->quantidade = $request->quantidade;
-    $distribuicao_item->quantidade_falta = $request->quantidade_falta;
-    $distribuicao_item->quantidade_danificados = $request->quantidade_danificados;
-    $distribuicao_item->item_id = $request->item_id;
-    $distribuicao_item->distribuicao_id = $request->distribuicao_id;
+    $distribuicao_item = \App\Distribuicao_item::where('item_id','=',$request->item_id)
+                                               ->where('distribuicao_id','=',$request->distribuicao_id) 
+                                               ->get()->first();
 
-    $distribuicao_item->save();
-
-    $itens = \App\Item::all();
     $distribuicao = \App\Distribuicao::find($request->distribuicao_id);
-    session()->flash('success', 'Item adicionado.');
+
+    if (!isset($distribuicao_item)) {
+      if (isset($distribuicao)) {
+        $distribuicao_item = new \App\Distribuicao_item();
+        $distribuicao_item->quantidade = $request->quantidade;
+        $distribuicao_item->quantidade_falta = $request->quantidade_falta;
+        $distribuicao_item->quantidade_danificados = $request->quantidade_danificados;
+        $distribuicao_item->item_id = $request->item_id;
+        $distribuicao_item->distribuicao_id = $request->distribuicao_id;
+        $distribuicao_item->save();
+  
+        $itens = \App\Item::all();
+        session()->flash('success', 'Item adicionado.');
+        return view("InserirItensDistribuicao", ["distribuicao" => $distribuicao, "itens" => $itens]);
+      }
+      
+      session()->flash('success', 'Distribuicao não existe.');
+      return redirect()->route('/distribuicao/listar');
+    }
+    $itens = \App\Item::all();
+    session()->flash('success', 'Esse item já existe.');
     return view("InserirItensDistribuicao", ["distribuicao" => $distribuicao, "itens" => $itens]);
   }
 
   public function removerItemDistribuicao(Request $request) {
     $distribuicao_item = \App\Distribuicao_item::find($request->id);
     $itens = \App\Item::all();
+    
+
+    if (isset($distribuicao_item)) {
+      $distribuicao = \App\Distribuicao::find($distribuicao_item->distribuicao_id);
+
+      if (isset($distribuicao)) {
+        $distribuicao_item->delete();
+
+        session()->flash('success', 'Item removido.');
+        return view("InserirItensDistribuicao", ["distribuicao" => $distribuicao, "itens" => $itens]);
+      }
+      
+      session()->flash('success', 'Distribuicao não existe.');
+      return redirect()->route('/distribuicao/listar');
+    }
     $distribuicao = \App\Distribuicao::find($distribuicao_item->distribuicao_id);
-
-    $distribuicao_item->delete();
-
-    session()->flash('success', 'Item removido.');
+    session()->flash('success', 'Esse Item não existe na Distribuição.');
     return view("InserirItensDistribuicao", ["distribuicao" => $distribuicao, "itens" => $itens]);
   }
